Fail dev-dependencies if the child fails
This fix is a bit iffy, and I need to go back and clean it up better.
BUT what it's doing is checking if the child process hasn't responded
to the health check at all for 20 times (with a 100ms between checks,
that makes 2s) in a row, then exit because it's probably dead.

I'm pretty sure the extra error channel and ctx.Done() stuff I did is
overkill and I can remove it. But the better part would be to rethink
how this program works and redo the entire part. Instead of doing that,
I will ignore it for now and hope that I'll deal with it better in the
future 😅