Switch to go-chi for routing
I struggled to configure the router with /{id}-{slug} as my path and I
had already decided to switch to Chi to make using middlewares easier.

So when I struggled for ~15-20m to get the routing to work I tried out
Chi and it worked nearly immediately, and then just added some
recommended basic middlewares and went from there.