Pull down the Postgres container on bootstrap
It's kinda big, so not having to pull it down during the test run is
nice, so we don't hit timeouts from pulling containers.

Did this because I thought some container problems with Colima was
because getting containers were slow. It wasn't, but decided to keep it
because it's still nice when I had figured out the container to make
the wrapping was easy.