Bind contributing causes with a why
So we can say that this incident was because of unbounded
resource usage, and the reason we say that is because we didn't have a
limit on how many go routines we created, so we ran out of memory with
all waiting routines.

This also meant we needed to create all the required contributing
causes, or a method for storing and retrieving them, so we can
reference or normalized cases.

I'm not super keen on the package normalized, but I've also been
talking about them as my "normalized contributing causes" and
"normalized triggers," so… it fits what I'm calling things.