Do validation in the review service and storage
The storage should only be concerned with persistence and no extra
logic. And it also means I need to move the created/updated at out.

It kinda implies I should not be using auto-incrementing IDs because
I'm relying on the storage to decide for me. It should be in the
service, too.