Move the timestamp updates to the service layer
I'm not super keen on how this works out in practice, because I feel
like I'm validating business logic in the service layer. And I'm doing
that because the only way I could verify the mock call would be by
making reviewing.Review an interface so I could pass in a fake
implementation, but then I lose out on so much else. It doesn't feel
entirely kosher.

I should go read the DDD in Go book and see if I can get inspiration
from how it deals with this concern.