Move logic out of service and into a collaborator
So that the tests in the service are way more reasonable and so as I
keep extending the logic they fit together better.

I kinda like where this is going, but I feel a bit iffy about it.
Looking forward to hearing more from other people as I show it.